Fix Some Errors

+++ Http/Controllers/destroy.php
@@ -7,10 +7,12 @@ $db = App::resolve(Database::class);
 
 if (isset($_POST['delete'])) {
     $ids = $_POST['delete'];
-    foreach ($ids as $id) {
-        $db->query('DELETE FROM products WHERE id = :id', [
-            'id' => $id
-        ]);
+    if (is_array($ids)) {
+        foreach ($ids as $id) {
+            $db->query('DELETE FROM products WHERE id = :id', [
+                'id' => $id
+            ]);
+        }
     }
 }
